travis to only push to gh-pages when branch is master

+++ deploy.sh
@@ -1,17 +1,23 @@
-cd docs
-git init
+#!/bin/bash
 
-# inside this git repo create Travis user
-git config user.name "Travis CI"
-git config user.email "james@jamesbirrellgray.com"
+BRANCH_NAME=$(git rev-parse --symbolic-full-name --abbrev-ref HEAD)
 
-# The first and only commit to this new Git repo contains all the
-# files present with the commit message "Deploy to GitHub Pages".
-git add .
-git commit -m "Deploy to GitHub Pages"
+if [ $BRANCH_NAME = "master" ]; then
+  cd docs
+  git init
 
-# Force push from the current repo's master branch to the remote
-# repo's gh-pages branch. (All previous history on the gh-pages branch
-# will be lost, since we are overwriting it.) We redirect any output to
-# /dev/null to hide any sensitive credential data that might otherwise be exposed.
-git push --force --quiet "https://${GH_TOKEN}@${GH_REF}" master:gh-pages > /dev/null 2>&1
+  # inside this git repo create Travis user
+  git config user.name "Travis CI"
+  git config user.email "james@jamesbirrellgray.com"
+
+  # The first and only commit to this new Git repo contains all the
+  # files present with the commit message "Deploy to GitHub Pages".
+  git add .
+  git commit -m "Deploy to GitHub Pages"
+
+  # Force push from the current repo's master branch to the remote
+  # repo's gh-pages branch. (All previous history on the gh-pages branch
+  # will be lost, since we are overwriting it.) We redirect any output to
+  # /dev/null to hide any sensitive credential data that might otherwise be exposed.
+  git push --force --quiet "https://${GH_TOKEN}@${GH_REF}" master:gh-pages > /dev/null 2>&1
+fi
